feat: put getDevRoute and getCommand in build-utils
