+++ packages/now-static-build/src/index.ts
@@ -1,7 +1,4 @@
 import path from 'path';
-import spawn from 'cross-spawn';
-import getPort from 'get-port';
-import { timeout } from 'promise-timeout';
 import { existsSync, readFileSync, statSync, readdirSync } from 'fs';
 import frameworks from './frameworks';
 import {
@@ -16,19 +13,9 @@ import {
   Route,
   BuildOptions,
   Config,
+  getCommand,
+  getDevRoute,
 } from '@now/build-utils';
-
-interface PackageJson {
-  scripts?: {
-    [key: string]: string;
-  };
-  dependencies?: {
-    [key: string]: string;
-  };
-  devDependencies?: {
-    [key: string]: string;
-  };
-}
 
 interface Framework {
   name: string;
@@ -67,52 +54,11 @@ function validateDistDir(distDir: string, isDev: boolean | undefined) {
   }
 }
 
-function getCommand(pkg: PackageJson, cmd: string, config: Config) {
-  // The `dev` script can be `now dev`
-  const nowCmd = `now-${cmd}`;
-  const { zeroConfig } = config;
-
-  if (!zeroConfig && cmd === 'dev') {
-    return nowCmd;
-  }
-
-  const scripts = (pkg && pkg.scripts) || {};
-
-  if (scripts[nowCmd]) {
-    return nowCmd;
-  }
-
-  if (scripts[cmd]) {
-    return cmd;
-  }
-
-  return nowCmd;
-}
-
 export const version = 2;
 
-const nowDevScriptPorts = new Map();
+export async function build(buildOptions: BuildOptions) {
+  const { files, entrypoint, workPath, config, meta = {} } = buildOptions;
 
-const getDevRoute = (srcBase: string, devPort: number, route: Route) => {
-  const basic: Route = {
-    src: `${srcBase}${route.src}`,
-    dest: `http://localhost:${devPort}${route.dest}`,
-  };
-
-  if (route.headers) {
-    basic.headers = route.headers;
-  }
-
-  return basic;
-};
-
-export async function build({
-  files,
-  entrypoint,
-  workPath,
-  config,
-  meta = {},
-}: BuildOptions) {
   console.log('Downloading user files...');
   await download(files, workPath, meta);
 
@@ -175,92 +121,14 @@ export async function build({
     await runNpmInstall(entrypointDir, ['--prefer-offline'], spawnOpts);
 
     if (meta.isDev && pkg.scripts && pkg.scripts[devScript]) {
-      let devPort: number | undefined = nowDevScriptPorts.get(entrypoint);
-
       if (framework && framework.defaultRoutes) {
         // We need to delete the routes for `now dev`
         // since in this case it will get proxied to
-        // a custom server we don't have controll over
+        // a custom server we don't have control over
         delete framework.defaultRoutes;
       }
 
-      if (typeof devPort === 'number') {
-        console.log(
-          '`%s` server already running for %j',
-          devScript,
-          entrypoint
-        );
-      } else {
-        // Run the `now-dev` or `dev` script out-of-bounds, since it is assumed that
-        // it will launch a dev server that never "completes"
-        devPort = await getPort();
-        nowDevScriptPorts.set(entrypoint, devPort);
-
-        const opts = {
-          cwd: entrypointDir,
-          env: { ...process.env, PORT: String(devPort) },
-        };
-
-        const child = spawn('yarn', ['run', devScript], opts);
-        child.on('exit', () => nowDevScriptPorts.delete(entrypoint));
-        if (child.stdout) {
-          child.stdout.setEncoding('utf8');
-          child.stdout.pipe(process.stdout);
-        }
-        if (child.stderr) {
-          child.stderr.setEncoding('utf8');
-          child.stderr.pipe(process.stderr);
-        }
-
-        // Now wait for the server to have listened on `$PORT`, after which we
-        // will ProxyPass any requests to that development server that come in
-        // for this builder.
-        try {
-          await timeout(
-            new Promise(resolve => {
-              const checkForPort = (data: string) => {
-                // Check the logs for the URL being printed with the port number
-                // (i.e. `http://localhost:47521`).
-                if (data.indexOf(`:${devPort}`) !== -1) {
-                  resolve();
-                }
-              };
-              if (child.stdout) {
-                child.stdout.on('data', checkForPort);
-              }
-              if (child.stderr) {
-                child.stderr.on('data', checkForPort);
-              }
-            }),
-            5 * 60 * 1000
-          );
-        } catch (err) {
-          throw new Error(
-            `Failed to detect a server running on port ${devPort}.\nDetails: https://err.sh/zeit/now-builders/now-static-build-failed-to-detect-a-server`
-          );
-        }
-
-        console.log('Detected dev server for %j', entrypoint);
-      }
-
-      let srcBase = mountpoint.replace(/^\.\/?/, '');
-
-      if (srcBase.length > 0) {
-        srcBase = `/${srcBase}`;
-      }
-
-      if (framework && framework.defaultRoutes) {
-        for (const route of framework.defaultRoutes) {
-          routes.push(getDevRoute(srcBase, devPort, route));
-        }
-      }
-
-      routes.push(
-        getDevRoute(srcBase, devPort, {
-          src: '/(.*)',
-          dest: '/$1',
-        })
-      );
+      routes.push(await getDevRoute(buildOptions));
     } else {
       if (meta.isDev) {
         console.log(`WARN: "${devScript}" script is missing from package.json`);
